move stuff into a db directory
